backend/requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-PyMongo
pymongo
python-dotenv
flask-cors
